Refactor(App, Race, utils, Validator, View): refactor code structure
- Create View class for console I/O operations
  - Move readCarNames(), readAttemptCount(), printProgress(), printResult()
- Create Race class for racing game logic
  - Move execute(), executeAttempt()
- Create utils file for utility functions
  - Move parseCarNames(), calculateResult()
