+++ src/App.js
@@ -1,86 +1,17 @@
-import { Console, Random } from "@woowacourse/mission-utils";
-import Car from "./Car.js";
-import Validator from "./Validator.js";
+import { Console } from "@woowacourse/mission-utils";
+import { calculateResult } from "./utils.js";
+
+import View from "./View.js";
+import Race from "./Race.js";
 
 class App {
-  parseCarNames(userCarNames) {
-    const carNames = userCarNames.split(",").filter((car) => car.trim() !== "");
-    Validator.validateCarNamesInput(carNames);
-    return carNames;
-  }
-  async readCarNames(cars) {
-    let userCarNames = "";
-
-    while (!userCarNames.trim()) {
-      userCarNames = await Console.readLineAsync(
-        "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)\n"
-      );
-    }
-
-    const parsedCarNames = this.parseCarNames(userCarNames);
-
-    parsedCarNames.forEach((car) => cars.push(new Car(car)));
-
-    return cars;
-  }
-
-  async readAttemptCount() {
-    let userAttemptCount = 0;
-    while (userAttemptCount === 0) {
-      userAttemptCount = Number(
-        await Console.readLineAsync("시도할 횟수는 몇 회인가요?\n")
-      );
-    }
-    Validator.validateAttemptCountInput(userAttemptCount);
-    return userAttemptCount;
-  }
-
-  excuteAttempt(cars) {
-    cars.forEach((car) => {
-      const randomNuber = Random.pickNumberInRange(0, 9);
-      if (randomNuber > 3) {
-        car.increaseDistance();
-      }
-    });
-  }
-
-  printProgress(cars) {
-    cars.forEach((car) => {
-      Console.print(`${car.name} : ${"-".repeat(car.distance)}`);
-    });
-    Console.print(`\n`);
-  }
-
-  runRace(cars, attemptCount) {
-    Console.print("\n실행 결과");
-    for (let i = 0; i < attemptCount; i++) {
-      this.excuteAttempt(cars);
-      this.printProgress(cars);
-    }
-  }
-
-  calculateResult(cars) {
-    const sortedCars = [...cars].sort(
-      (prev, next) => next.distance - prev.distance
-    );
-    const result = sortedCars.filter(
-      (car) => car.distance === sortedCars[0].distance
-    );
-    return result;
-  }
-
-  printResult(result) {
-    const resultString = result.map((car) => car.name).join(", ");
-    Console.print(`최종 우승자 : ${resultString}`);
-  }
-
   async run() {
-    const cars = [];
-    await this.readCarNames(cars);
-    const attemptCount = await this.readAttemptCount();
-    this.runRace(cars, attemptCount);
-    const result = this.calculateResult(cars);
-    this.printResult(result);
+    const cars = await View.readCarNames();
+    const attemptCount = await View.readAttemptCount();
+    Console.print("\n실행 결과");
+    Race.excute(cars, attemptCount);
+    const result = calculateResult(cars);
+    View.printResult(result);
   }
 }
